// src/store/searchdetail/search-type.ts
export interface searchdetailtype {
  searchdetail: any[]
  searchkeywords: any
  searchSinger: any
  searchAlbum: any // 获取搜索的专辑
  searchMV: any // 获取搜索的视频
  searchSong: any // 获取搜索的歌单
  searchLyrics: any // 获取搜索的歌词
  searchUser: any // 获取搜索的用户
  userevents: any
}